Add files via upload
The project activity is a bundled Solar PV installation of total capacity 2.415 MW (2415.2 kWp), located in various locations of the two states namely Tamil Nādu and Maharashtra, of India. The promoter of the project activity is “SolFreedom Power Limited” (hereinafter also referred to as SFPL), who has developed and been operating the SPV plants. There are fourteen different SPV plants included under this bundle which are installed in the industrial/commercial premises of the respective power users (i.e. referred to as the consumers or bundle members). SFPL is a solar EPC and service provider, has executed individual power purchase agreement with all the bundle
members and further to execute the process of submission of the total capacity as bundled project under “Universal Carbon Registry” (UCR). The SPV plants are operational with continuity and reduction of Greenhouse Gas emissions (GHGs), currently being applied under “Universal Carbon Registry” (UCR) for receiving carbon finance for a long term sustainability of the project.